requirements.txt
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
bokeh==1.0.4
colorama==0.4.1
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.14.0
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.5.0
jupyterlab==0.35.6
jupyterlab-server==0.2.0
llvmlite==0.29.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
nbserverproxy==0.8.8
nodejs==0.1.1
notebook==5.7.8
numba==0.43.1
numpy==1.16.4
optional-django==0.1.0
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.0
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
Pygments==2.4.2
pyparsing==2.4.0
pyrsistent==0.15.2
python-dateutil==2.8.0
pytz==2019.1
pywinpty==0.5.5
PyYAML==5.1.1
pyzmq==18.0.1
scikit-learn==0.20.3
scipy==1.3.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
umap-learn==0.3.8
wcwidth==0.1.7
webencodings==0.5.1
